Updated Project files and reduced warnings
Updated Project files to exclude 32 bit systems, added cast to main cpp file to avoid warning

+++ LinkErrorHelper/LinkErrorHelperTool/LinkErrorHelperTool.cpp
@@ -38,7 +38,7 @@ namespace /*anonymous*/ {
         //throw std::runtime_error("failed to run command");
     }
 
-    while (fgets(buffer.data(), buffer.size(), pipe.get()) != nullptr) 
+    while (fgets(buffer.data(), static_cast<int>(buffer.size()), pipe.get()) != nullptr) 
     {
         result << buffer.data();
     }
